+++ BlowinCleanCode/BlowinCleanCode/Feature/CodeSmell/CatchShouldDoMoreThanRethrowFeatureAnalyze.cs
@@ -1,4 +1,7 @@
-﻿using BlowinCleanCode.Feature.Base;
+﻿using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+using BlowinCleanCode.Feature.Base;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -6,7 +9,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace BlowinCleanCode.Feature.CodeSmell
 {
-    public sealed class CatchShouldDoMoreThanRethrowFeatureAnalyze : FeatureSyntaxNodeAnalyzerBase<CatchClauseSyntax>
+    public sealed class CatchShouldDoMoreThanRethrowFeatureAnalyze : FeatureSyntaxNodeAnalyzerBase<TryStatementSyntax>
     {
         public override DiagnosticDescriptor DiagnosticDescriptor { get; } = new DiagnosticDescriptor(Constant.Id.CatchShouldDoMoreThanRethrow, 
             title: "Catch should do more than rethrow",
@@ -15,26 +18,60 @@ namespace BlowinCleanCode.Feature.CodeSmell
             DiagnosticSeverity.Warning, 
             isEnabledByDefault: true);
         
-        protected override SyntaxKind SyntaxKind => SyntaxKind.CatchClause;
+        protected override SyntaxKind SyntaxKind => SyntaxKind.TryStatement;
         
-        protected override void Analyze(SyntaxNodeAnalysisContext context, CatchClauseSyntax syntaxNode)
+        protected override void Analyze(SyntaxNodeAnalysisContext context, TryStatementSyntax tryStatementSyntax)
         {
-            var statements = syntaxNode.Block?.Statements;
-            if(statements == null)
+            var locations = GetInvalidLocations(tryStatementSyntax);
+            if(locations.IsEmpty)
                 return;
+
+            ReportDiagnostic(context, locations[0], locations.Skip(1));
+        }
+
+        private ImmutableArray<Location> GetInvalidLocations(TryStatementSyntax tryStatementSyntax)
+        {
+            if(!ShouldReport(tryStatementSyntax))
+                return ImmutableArray<Location>.Empty;
+            
+            var builder = ImmutableArray.CreateBuilder<Location>();
+            foreach (var clauseSyntax in tryStatementSyntax.Catches)
+            {
+                var throwStatement = GetThrowStatementSyntax(clauseSyntax);
+                builder.Add(throwStatement.GetLocation());
+            }
+
+            return builder.ToImmutable();
+        }
+
+        private bool ShouldReport(TryStatementSyntax tryStatementSyntax)
+        {
+            foreach (var catchClauseSyntax in tryStatementSyntax.Catches)
+            {
+                var throwStatement = GetThrowStatementSyntax(catchClauseSyntax);
+                if(throwStatement == null || AnalyzerCommentSkipCheck.Skip(throwStatement))
+                    return false;
+
+                if (!IsInvalid(throwStatement, catchClauseSyntax))
+                    return false;
+            }
+
+            return true;
+        }
+
+        private static ThrowStatementSyntax GetThrowStatementSyntax(CatchClauseSyntax catchClauseSyntax)
+        {
+            var statements = catchClauseSyntax.Block?.Statements;
+            if(statements == null)
+                return null;
 
             var unwrapStatements = statements.Value;
             if(unwrapStatements.Count != 1)
-                return;
+                return null;
 
-            var throwStatement = unwrapStatements.First() as ThrowStatementSyntax;
-            if(throwStatement == null || AnalyzerCommentSkipCheck.Skip(throwStatement))
-                return;
-            
-            if(IsInvalid(throwStatement, syntaxNode))
-                ReportDiagnostic(context, throwStatement.GetLocation());
+            return unwrapStatements.First() as ThrowStatementSyntax;
         }
-
+        
         private static bool IsInvalid(ThrowStatementSyntax throwExpressionSyntax, CatchClauseSyntax syntaxNode)
         {
             switch (throwExpressionSyntax.Expression)